sit update mds

+++ application/modules/mds_set_measure/controllers/mds_set_measure.php
@@ -6,7 +6,7 @@ Class Mds_set_measure extends  Mdevsys_Controller{
 	function __construct(){
 		parent::__construct();
 		$this->load->model('mds_set_measure/Mds_set_measure_model','measure');
-		
+		$this->load->model('mds_set_indicator/Mds_set_metrics_model','metrics');
 	}
 	
 	public $urlpage = "mds_set_measure";
@@ -61,9 +61,19 @@ Class Mds_set_measure extends  Mdevsys_Controller{
 		$urlpage = $this->urlpage;
 		if(!is_login())redirect("home");
 		if(is_permit(login_data('id'),1) == '')redirect("mds"); // ตรวจสอบว่าเป็น กพร. หรือไม่		
-	
-		new_save_logfile("DELETE",$this->modules_title,$this->measure->table,"ID",$ID,"measure_name",$this->modules_name);					
-		$this->measure->delete($ID);		
+		if($ID != ''){
+				$chk_metrics = "select * from mds_set_metrics where mds_set_measure_id = '".$ID."'";
+				$result_chk_metrics =  $this->metrics->get($chk_metrics);
+				$num_chk = count($result_chk_metrics);
+			if($num_chk == '0'){
+				new_save_logfile("DELETE",$this->modules_title,$this->measure->table,"ID",$ID,"measure_name",$this->modules_name);					
+				$this->measure->delete($ID);
+				set_notify('error', 'ลบข้อมูลเสร็จสิน');	
+			}else{
+			    set_notify('error', "ไม่สามารถลบหน่วยวัดได้ เนื่องจากมีการใช้ชื่อหน่วยวัดนี้อยู่");
+			}
+		}
+			
 		redirect($urlpage);
 	}
 	function check_measure_name(){
